Move the im::encoding types to a new im::encoding module
