+++ rs-matter/src/im/encoding/attr/subscribe_builder.rs
@@ -18,7 +18,7 @@
 //! Streaming TLV builder for `SubscribeRequestMessage`.
 //!
 //! Same typestate-machine + implicit-skip convention as
-//! [`crate::im::attr::read_builder::ReadReqBuilder`]: mandatory
+//! [`crate::im::encoding::attr::read_builder::ReadReqBuilder`]: mandatory
 //! fields must be written in order (`keep_subs`, `min_int_floor`,
 //! `max_int_ceil`, `fabric_filtered`); optional intermediate fields
 //! (`AttributeRequests`, `EventRequests`, `EventFilters`,
@@ -26,7 +26,7 @@
 //! setter — later-field setters are available on earlier states so
 //! the user can write a minimal request in one straight chain. The
 //! `AttrPath` array sub-builder is re-used from
-//! [`crate::im::attr::read_builder`] since the wire shape is
+//! [`crate::im::encoding::attr::read_builder`] since the wire shape is
 //! identical to the read variant.
 //!
 //! # Layout
