+++ rs-matter/src/im/encoding/attr/write.rs
