+++ rs-matter/src/im/encoding/attr/write_builder.rs
